rt: add note about softirq delta found from sanity check
Rather than re-export all the rt patches from -rt80 and
create a lot of meaningless git noise, it was decided to
instead do a sanity check as follows:

1) take -rt's tag v3.10.75-rt80
2) Merge v3.10.76 and 77, ... v3.10.80
3) Layer on the small number of patches in our stable/ dir
4) Compare to our treegen automatic tag v3.10.80-rt80

After doing that there are three deltas.  The 1st is a local
change to our .gitignore done during treegen (cosmetic).
The 2nd is also largely cosmetic; an extra set of braces
added to a #ifdef statement in drivers/gpu/drm/i915/i915_gem.c

What matters is the 3rd change:

   --- a/kernel/softirq.c
   +++ b/kernel/softirq.c
   @@ -210,7 +210,7 @@ static void handle_pending_softirqs(u32 pending, int cpu, int need_rcu_bh_qs)
    static void run_ksoftirqd(unsigned int cpu)
    {
           local_irq_disable();
   -       if (ksoftirqd_softirq_pending()) {
   +       if (local_softirq_pending()) {

This change appears as unintentional in the upstream feed so,
we document its existence here in the scc file but do not
make the change locally, as it does not exist in any other
rt version that was comparable enough to be checked.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>