vfs: read file_handle only once in handle_to_path - linux - CVE-2015-1420
Issue: LIN6-9645

1/1 [
Author: Sasha Levin
Email: sasha.levin@oracle.com
Subject: vfs: read file_handle only once in handle_to_path
Date: Wed, 28 Jan 2015 15:30:43 -0500

commit 161f873b89136eb1e69477c847d5a5033239d9ba upstream

We used to read file_handle twice.  Once to get the amount of extra
bytes, and once to fetch the entire structure.

This may be problematic since we do size verifications only after the
first read, so if the number of extra bytes changes in userspace between
the first and second calls, we'll have an incoherent view of
file_handle.

Instead, read the constant size once, and copy that over to the final
structure without having to re-read it again.

Signed-off-by: Sasha Levin <sasha.levin@oracle.com>
Cc: Al Viro <viro@zeniv.linux.org.uk>
Cc: stable@vger.kernel.org
Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Hu <yadi.hu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>