esdhc: Workaround for eSDHC clock glitch issue
Issue: LIN4-32700

1/3 [
Author: Philip Rakity
Email: prakity@marvell.com
Subject: mmc: sdhci: Add pre and post reset processing for chip specific reset
Date: Fri, 21 Jan 2011 11:26:40 -0800

commit 393c1a34dd852dbf0c409ec94dd48057100480ea upstream

Marvell pxa controllers have private registers that may need to be
modified before and after a reset is done.

For example, the SD reset operation, RESET_ALL, will reset the private
registers to their default state.  This will cause the clock adjustment
registers that may have been programmed to have incorrect values.

RESET_DATA sometimes needs to be delayed before the reset is done
(depending on SoC) to enable any transactions being handled by the
SDIO card to be completed.  Needed in pre SD 3.0 silicon to handle
clock gating.

Implement hooks to allow this to happen.

Signed-off-by: Philip Rakity <prakity@marvell.com>
Signed-off-by: Chris Ball <cjb@laptop.org>
Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

2/3 [
Author: "Haijun.Zhang"
Email: haijun.zhang@freescale.com
Subject: ESDHC-powerpc mmc:host workaround for glitch generated when clock changed
Date: Fri, 22 Mar 2013 13:57:23 +0800

A-005055: SDHC: Glitch is generated on the card clock with software reset or
clock divider change.

A simple workaround is to disable the SD card clock before the software reset,
and enable it when the module resumes normal operation.The Host and the SD card
are in a master-slave relationship. The Host provides clock and control
transfer across the interface. Therefore, any existing operation is discarded
when the Host controller is reset.
The recommended flow is as follows:
1. Software disable bit[3], SDCLKEN, of the System Control Register
2. Trigger software reset and/or set clock divider
3. Check bit[3], SDSTB, of the Present State Register for stable clock
4. Enable bit[3], SDCLKEN, of the System Control Register
Using the above method, the eSDHC cannot send command or transfer data when
there is a glitch in the clock line, and the glitch does not cause any issue.

In case the reason of Reset All command, we use reset data and reset command
to instead.
In case the reason of Clock change, this situation can only be occurred in card
setup phases, so just go idle state and re-init the card(Card detect can do
this for us).

Signed-off-by: Jerry Huang <Chang-Ming.Huang@freescale.com>
Signed-off-by: Haijun Zhang <haijun.zhang@freescale.com>
Change-Id: I1add2738ecf77369adc9f30718bd6a6e08ee4f17
Reviewed-on: http://git.am.freescale.net:8181/569
Reviewed-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
Tested-by: Fleming Andrew-AFLEMING <AFLEMING@freescale.com>
[Original patch taken from QorIQ-SDK-V1.4-SOURCE-20130625-yocto.iso
 Change SDHCI_QUIRK2_BROKEN_RESET_ALL to (1<<4) to port 3.10 kernel]
Signed-off-by: Bin Jiang <bin.jiang@windriver.com>
[ o use host->quirks instead of host->quirks2 which is not avaiable
	in the current kernel
  o discard the code specific to the sdhci-pltfm.[ch], which is not
	 applicable to the current kernel
]
Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

3/3 [
Author: Haijun Zhang
Email: Haijun.Zhang@freescale.com
Subject: esdhc: Workaround for eSDHC clock glitch issue
Date: Mon, 14 Oct 2013 16:37:38 +0800

A-003980: SDHC: Glitch is generated on the card clock with software reset or
clock divider change
Description: A glitch may occur on the SDHC card clock when the software sets
the RSTA bit (software reset) in the system control register. It can also be
generated by setting the clock divider value. The glitch produced can cause
the external card to switch to an unknown state. The occurrence is not
deterministic.
Workaround:
A simple workaround is to disable the SD card clock before the software reset,
and enable it when the module resumes normal operation.
The Host and the SD card are in a master-slave relationship. The Host provides
clock and control transfer across the interface. Therefore, any existing
operation is discarded when the Host controller is reset.
The recommended flow is as follows:
1. Software disable bit[3], SDCLKEN, of the System Control Register
2. Trigger software reset and/or set clock divider
3. Check bit[3], SDSTB, of the Present State Register for stable clock
4. Enable bit[3], SDCLKEN, of the System Control Register
Using the above method, the eSDHC cannot send command or transfer data when

there is a glitch in the clock line, and the glitch does not cause any issue.
Signed-off-by: Haijun Zhang <haijun.zhang@freescale.com>
Change-Id: I1d8c3757ffdab73aff92cf1f0a6b89f82d11acfe
Reviewed-on: http://git.am.freescale.net:8181/4357
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Xie Xiaobo-R63061 <X.Xie@freescale.com>
Reviewed-by: Rivera Jose-B46482 <German.Rivera@freescale.com>
[Original patch taken from QorIQ-SDK-V1.5-SOURCE-20131219-yocto.iso
 Change SDHCI_QUIRK2_DISABLE_CLOCK_BEFORE_RESET to (1<<8) to port 3.10
 kernel]
Signed-off-by: Bin Jiang <bin.jiang@windriver.com>
[ add "fsl,p1020-esdhc" for device compatibility checking]

Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>