e1000e: Fix Hardware Unit Hang
1/1 [
Author: David Ertman
Email: davidx.m.ertman@intel.com
Subject: e1000e: Fix Hardware Unit Hang
Date: Wed, 8 Jan 2014 01:07:55 +0000

commit d9554e9659e974904fff3e93b2833f67e87ec084 upstream

The check for pending Tx work when link is lost was mistakenly moved to be
done only when link is first detected to be lost.  It turns out there is a
small window of opportunity for additional Tx work to get queued up shortly
after link is dropped.

Move the check back to the place it was before in the watchdog task.  Put in
additional debug information for other reset paths and a final catch-all for
false hangs in the scheduled function that prints out the hardware hang
message.

Signed-off-by: Dave Ertman <davidx.m.ertman@intel.com>
Signed-off-by: Bruce Allan <bruce.w.allan@intel.com>
Tested-by: Jeff Pieper <jeffrey.e.pieper@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>