ixgbe: modify behavior on receiving a HW ECC error
Issue: LIN6-10121

1/1 [
Author: Don Skidmore
Email: donald.c.skidmore@intel.com
Subject: ixgbe: modify behavior on receiving a HW ECC error.
Date: Tue, 25 Feb 2014 17:58:53 -0800

commit d773ce2de1c670e0d259870a2ea8fd9f60ab98cd upstream.

Currently when we noticed a HW ECC error we would request the use reload
the driver to force a reset of the part.  This was done due to the mistaken
believe that a normal reset would not be sufficient.  Well it turns out it
would be so now we just schedule a reset upon seeing the ECC.

Signed-off-by: Don Skidmore <donald.c.skidmore@intel.com>
Tested-by: Phil Schmitt <phillip.j.schmitt@intel.com>
Signed-off-by: Aaron Brown <aaron.f.brown@intel.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: xiao zhang <xzhang1@ala-lpggp1.wrs.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>