powerpc: Fix a wrong sys_call_table declaration to enable syscall tracing
Issue: LIN7-4198

1/1 [
Author: Zumeng Chen
Email: zumeng.chen@windriver.com
Subject: powerpc: Fix a wrong sys_call_table declaration to enable syscall tracing
Date: Sat, 18 Jul 2015 10:53:20 +0800

The commit 155e149e included a very new mainline commit 1028ccf5 to fix a
wrong sys_call_table declaration without considerations on the dependency
between patches and arch difference. This patch is try to fix it because of
the following reason:

sys_call_table referred in C level is defined as a pointer to a pointer array
in assembler level for powerpc64, and it has been defined as an array, so fix
it to avoid syscall perf error.

Signed-off-by:Zumeng Chen <zumeng.chen@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>