net: allow netdev_all_upper_get_next_dev_rcu with rtnl lock held
Issue: LIN6-10098

1/1 [
Author: John Fastabend
Email: john.r.fastabend@intel.com
Subject: net: allow netdev_all_upper_get_next_dev_rcu with rtnl lock held
Date: Tue, 26 Nov 2013 06:33:52 +0000

commit 85328240c625f322af9f69c7b60e619717101d77 upstream.

It is useful to be able to walk all upper devices when bringing
a device online where the RTNL lock is held. In this case it
is safe to walk the all_adj_list because the RTNL lock is used
to protect the write side as well.

This patch adds a check to see if the rtnl lock is held before
throwing a warning in netdev_all_upper_get_next_dev_rcu().

Also because we now have a call site for lockdep_rtnl_is_held()
outside COFIG_LOCK_PROVING an inline definition returning 1 is
needed. Similar to the rcu_read_lock_is_held().

Fixes: 2a47fa45d4df ("ixgbe: enable l2 forwarding acceleration for macvlans")
CC: Veaceslav Falico <vfalico@redhat.com>
Reported-by: Yuanhan Liu <yuanhan.liu@linux.intel.com>
Signed-off-by: John Fastabend <john.r.fastabend@intel.com>
Tested-by: Phil Schmitt <phillip.j.schmitt@intel.com>
Signed-off-by: Jeff Kirsher <jeffrey.t.kirsher@intel.com>
Signed-off-by: Liwei Song <liwei.song@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>