dpa: set skb truesize to the skb length if jumbo is enabled
Issue: LIN5-20654

1/1 [
Author: yzhu1
Email: yanjun.zhu@windriver.com
Subject: dpa: set skb truesize to the skb length if jumbo is enabled
Date: Wed, 22 Jul 2015 15:14:19 +0800

This patch is from FSL SDK 1.6.

After the jubmo is set to 9600, the truesize of skb should be set
to the skb length before these skb packets are sent to netif_receive_skb.

Signed-off-by: yzhu1 <yanjun.zhu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>