MIPS: Simplify ptrace_getfpregs FPU IR retrieval
Issue: LIN7-4362

1/1 [
Author: Paul Burton
Email: paul.burton@imgtec.com
Subject: MIPS: Simplify ptrace_getfpregs FPU IR retrieval
Date: Fri, 24 Jul 2015 09:47:44 -0700

commit dab75dd956522ce19403c108f659ea9b339f2559 upstream

All architecturally defined bits in the FPU implementation register
are read only & unchanging. It contains some implementation-defined
bits but the architecture manual states "This bits are explicitly not
intended to be used for mode control functions" which seems to provide
justification for viewing the register as a whole as unchanging. This
being the case we can simply re-use the value we read at boot rather
than having to re-read it later, and avoid the complexity which that
read entails.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Cc: linux-mips@linux-mips.org
Patchwork: https://patchwork.linux-mips.org/patch/6147/
Signed-off-by: Ralf Baechle <ralf@linux-mips.org>
[Backported to 3.14 kernel]
Signed-off-by: Yang Shi <yang.shi@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>