acpi: disable selection of the aggregator driver
Issue: OVP-2505

Investigation has revealed that this driver is not aligned with the
core supported OVP use cases, nor is it functionally complete and ready
for production deployment.

Since it was just enabled as an attempt to provide all possible ACPI
modules without direct scrutiny at the individual level, we disable
it now that we know more about it.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>