MIPS: Simplify PTRACE_PEEKUSR for FPC_EIR
Issue: LIN7-4362

1/1 [
Author: Paul Burton
Email: paul.burton@imgtec.com
Subject: MIPS: Simplify PTRACE_PEEKUSR for FPC_EIR
Date: Fri, 24 Jul 2015 09:44:11 -0700

commit 3351047f01fe012abbb585b400d1c51b57ed011d upstream

All architecturally defined bits in the FPU implementation register
are read only & unchanging. It contains some implementation-defined
bits but the architecture manual states "This bits are explicitly not
intended to be used for mode control functions" which seems to provide
justification for viewing the register as a whole as unchanging. This
being the case we can simply re-use the value we read at boot rather
than having to re-read it later, and avoid the complexity which that
read entails.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>
Reviewed-by: Qais Yousef <qais.yousef@imgtec.com>
Cc: linux-mips@linux-mips.org
Patchwork: https://patchwork.linux-mips.org/patch/6144/
Signed-off-by: Ralf Baechle <ralf@linux-mips.org>
[Backported to 3.14 kernel]
Signed-off-by: Yang Shi <yang.shi@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>