fsl-ls20xx base patches
Rally: US58257

1/5 [
Author: Yanjiang Jin
Email: yanjiang.jin@windriver.com
Subject: irqchip: gicv3-its: Add missing part for "Fix ITT allocation"
Date: Wed, 29 Jul 2015 15:31:21 +0800

WRL has merged the upstream commit

c848126734e8621e81659d819922b20d93a2aa6d
    irqchip: gicv3-its: Fix ITT allocation

before, but omitted a few codes, add missing part.

Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>
]

2/5 [
Author: Marc Zyngier
Email: marc.zyngier@arm.com
Subject: irqchip: gicv3-its: Fix unsafe locking reported by lockdep
Date: Fri, 6 Mar 2015 16:37:43 +0000

commit 3e39e8f56c1c67cdd1e8f06da0d6b7c831818c76 upstream

When compiled with CONFIG_LOCKDEP, the kernel shouts badly, saying
that my locking is unsafe. I'm afraid the kernel is right:

       CPU0                    CPU1
       ----                    ----
  lock(&its->lock);
                               local_irq_disable();
                               lock(&irq_desc_lock_class);
                               lock(&its->lock);
  <Interrupt>
    lock(&irq_desc_lock_class);

 *** DEADLOCK ***

The fix is to always take its->lock with interrupts disabled.

Reported-by: Will Deacon <will.deacon@arm.com>
Signed-off-by: Marc Zyngier <marc.zyngier@arm.com>
Link: https://lkml.kernel.org/r/1425659870-11832-5-git-send-email-marc.zyngier@arm.com
Signed-off-by: Jason Cooper <jason@lakedaemon.net>
[Adjust line differences for different base.]
Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>
]

3/5 [
Author: Will Deacon
Email: will.deacon@arm.com
Subject: drivers/vfio: allow type-1 IOMMU instantiation on top of an ARM SMMU
Date: Fri, 14 Nov 2014 09:10:59 -0700

commit 5e9f36c59a4883c6d8db4a3c097e4fc2234742de upstream

The ARM SMMU driver is compatible with the notion of a type-1 IOMMU in
VFIO.

This patch allows VFIO_IOMMU_TYPE1 to be selected if ARM_SMMU=y.

Signed-off-by: Will Deacon <will.deacon@arm.com>
[aw: update for existing S390 patch]
Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
]

4/5 [
Author: Yanjiang Jin
Email: yanjiang.jin@windriver.com
Subject: Revert "iommu: Remove iommu_domain_has_cap() API function"
Date: Wed, 29 Jul 2015 13:54:43 +0800

This reverts commit ade124c857b09475afe41bf93572ec0cb397cfd9.

Previous commit enables VFIO_IOMMU_TYPE1 to compile
drivers/vfio/vfio_iommu_type1.c, it calls iommu_domain_has_cap(),
so revert commit ade124c857b to restore it.

Signed-off-by: Yanjiang Jin <yanjiang.jin@windriver.com>
]

5/5 [
Author: Alex Williamson
Email: alex.williamson@redhat.com
Subject: vfio/iommu_type1: Multi-IOMMU domain support
Date: Wed, 26 Feb 2014 11:38:36 -0700

commit 1ef3e2bc04223ff956dc62abaf2dff1f3322a431 upstream

We currently have a problem that we cannot support advanced features
of an IOMMU domain (ex. IOMMU_CACHE), because we have no guarantee
that those features will be supported by all of the hardware units
involved with the domain over its lifetime.  For instance, the Intel
VT-d architecture does not require that all DRHDs support snoop
control.  If we create a domain based on a device behind a DRHD that
does support snoop control and enable SNP support via the IOMMU_CACHE
mapping option, we cannot then add a device behind a DRHD which does
not support snoop control or we'll get reserved bit faults from the
SNP bit in the pagetables.  To add to the complexity, we can't know
the properties of a domain until a device is attached.

We could pass this problem off to userspace and require that a
separate vfio container be used, but we don't know how to handle page
accounting in that case.  How do we know that a page pinned in one
container is the same page as a different container and avoid double
billing the user for the page.

The solution is therefore to support multiple IOMMU domains per
container.  In the majority of cases, only one domain will be required
since hardware is typically consistent within a system.  However, this
provides us the ability to validate compatibility of domains and
support mixed environments where page table flags can be different
between domains.

To do this, our DMA tracking needs to change.  We currently try to
coalesce user mappings into as few tracking entries as possible.  The
problem then becomes that we lose granularity of user mappings.  We've
never guaranteed that a user is able to unmap at a finer granularity
than the original mapping, but we must honor the granularity of the
original mapping.  This coalescing code is therefore removed, allowing
only unmaps covering complete maps.  The change in accounting is
fairly small here, a typical QEMU VM will start out with roughly a
dozen entries, so it's arguable if this coalescing was ever needed.

We also move IOMMU domain creation to the point where a group is
attached to the container.  An interesting side-effect of this is that
we now have access to the device at the time of domain creation and
can probe the devices within the group to determine the bus_type.
This finally makes vfio_iommu_type1 completely device/bus agnostic.
In fact, each IOMMU domain can host devices on different buses managed
by different physical IOMMUs, and present a single DMA mapping
interface to the user.  When a new domain is created, mappings are
replayed to bring the IOMMU pagetables up to the state of the current
container.  And of course, DMA mapping and unmapping automatically
traverse all of the configured IOMMU domains.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Cc: Varun Sethi <Varun.Sethi@freescale.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>