Leap Second Isssue and Deadlock
Issue: LIN4-32844

1/4 [
Author: John Stultz
Email: john.stultz@linaro.org
Subject: time: Refactor accumulation of nsecs to secs
Date: Fri, 13 Jul 2012 01:21:54 -0400

commit 1f4f948706bcec1b51bf6492bf04057d2e21e273 upstream

We do the exact same logic moving nsecs to secs in the
timekeeper in multiple places, so condense this into a
single function.

Signed-off-by: John Stultz <john.stultz@linaro.org>
Reviewed-by: Ingo Molnar <mingo@kernel.org>
Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: Richard Cochran <richardcochran@gmail.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Link: http://lkml.kernel.org/r/1342156917-25092-6-git-send-email-john.stultz@linaro.org
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

2/4 [
Author: David Vrabel
Email: david.vrabel@citrix.com
Subject: timekeeping: Indicate that clock was set in the pvclock gtod notifier
Date: Thu, 27 Jun 2013 11:35:46 +0100

commit 780427f0e113b4c77dfff4d258c05a902cdb0eb9 upstream

If the clock was set (stepped), set the action parameter to functions
in the pvclock gtod notifier chain to non-zero.  This allows the
callee to only do work if the clock was stepped.

This will be used on Xen as the synchronization of the Xen wallclock
to the control domain's (dom0) system time will be done with this
notifier and updating on every timer tick is unnecessary and too
expensive.

Signed-off-by: David Vrabel <david.vrabel@citrix.com>
Cc: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Cc: John Stultz <john.stultz@linaro.org>
Cc: <xen-devel@lists.xen.org>
Link: http://lkml.kernel.org/r/1372329348-20841-4-git-send-email-david.vrabel@citrix.com
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

3/4 [
Author: John Stultz
Email: john.stultz@linaro.org
Subject: timekeeping: Fix potential lost pv notification of time change
Date: Wed, 11 Dec 2013 20:07:49 -0800

commit 5258d3f25c76f6ab86e9333abf97a55a877d3870 upstream

In 780427f0e11 (Indicate that clock was set in the pvclock
gtod notifier), logic was added to pass a CLOCK_WAS_SET
notification to the pvclock notifier chain.

While that patch added a action flag returned from
accumulate_nsecs_to_secs(), it only uses the returned value
in one location, and not in the logarithmic accumulation.

This means if a leap second triggered during the logarithmic
accumulation (which is most likely where it would happen),
the notification that the clock was set would not make it to
the pv notifiers.

This patch extends the logarithmic_accumulation pass down
that action flag so proper notification will occur.

This patch also changes the varialbe action -> clock_set
per Ingo's suggestion.

Cc: Sasha Levin <sasha.levin@oracle.com>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: David Vrabel <david.vrabel@citrix.com>
Cc: Konrad Rzeszutek Wilk <konrad.wilk@oracle.com>
Cc: Prarit Bhargava <prarit@redhat.com>
Cc: Richard Cochran <richardcochran@gmail.com>
Cc: <xen-devel@lists.xen.org>
Cc: stable <stable@vger.kernel.org> #3.11+
Signed-off-by: John Stultz <john.stultz@linaro.org>
Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

4/4 [
Author: Shan Hai
Email: shan.hai@windriver.com
Subject: timekeeping: call clock_was_set_delayed outside of xtime_lock protected area
Date: Thu, 16 Jul 2015 14:29:44 +0800

The xtime_lock will be obtained in the function call chain initiated
by clock_was_set_delayed() while the lock was already held, this is
a recursive locking we should avoid, this patch fixes this locking rule
violation by moving the clock_was_set_delayed() call out of the xtime_lock.

Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>