splice: Apply generic position and size checks to each write - linux CVE-2014-7822
Issue: SCP6-562

1/1 [
Author: Ben Hutchings
Email: ben@decadent.org.uk
Subject: splice: Apply generic position and size checks to each write
Date: Tue, 14 Jul 2015 17:27:33 +0800

commit 894c6350eaad7e613ae267504014a456e00a3e2a upstream(linux-stable)

We need to check the position and size of file writes against various
limits, using generic_write_check().  This was not being done for
the splice write path.  It was fixed upstream by commit 8d0207652cbe
("->splice_write() via ->write_iter()") but we can't apply that.

CVE-2014-7822

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>
[The patch comes from linux-stable, the reason is the patch on
linux-torvalds is big and complicated(it has many dependences),
so it's not aplicable on the version.]
Signed-off-by: Wenlin Kang <wenlin.kang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>