enable Display audio driver power well request for baytrail
Issue: LIN7-4402

1/2 [
Author: Mengdong Lin
Email: mengdong.lin@intel.com
Subject: ALSA: hda - add AZX_DCAPS_I915_POWERWELL to Baytrail
Date: Tue, 21 Apr 2015 13:12:23 +0800

commit 40cc2392f4b144197d05eec73c1560f42fc25def upstream.

This patch addes AZX_DCAPS_I915_POWERWELL to BYT (Baytrail).

Like Braswell and Skylake, the HDMI codec on Bytrail is also in the shared
power well with GPU. This power well must be turned on before we reset link
to probe the codec, to avoid communication failure with the codec.

The side effect is that this power is always ON in S0 because the BYT HDMI
codec does not support EPSS or D3ClkStop and so the controller doesn't enter
D3 at runtime, and the HDMI codec and analog codec share a single physical
HD-A link and so we cannot reset the HD-A link freely when we re-enable the
power to use the HDMI codec.

Next step is to test if an AGP reset or double AGP reset on BYT HDMI codec is
okay to bring the HDMI codec back to a functional state after restoring the
power. If okay, we can bind the power on/off with the HDMI codec PM without
interrupting the analog audio.

Signed-off-by: Mengdong Lin <mengdong.lin@intel.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>
Signed-off-by: Liwei Song <liwei.song@windriver.com>
]

2/2 [
Author: Liwei Song
Email: liwei.song@windriver.com
Subject: enable Display audio driver power well request for baytrail
Date: Sun, 30 Aug 2015 17:49:37 +0800

This patch is use to enable Display audio driver in baytail platform,
It use the same power domain method API as Haswell and Broadwell.
Since HDMI codec on Bytrail is also in the shared power well with GPU,
this is use to avoid communication failure with the codec.

Meanwhile, it avoid failure when HDMI/VGA hotplug, because when hotplug
happened, all the other display of domains was disabled, without this
audio display, the communication with codec will fail.

Upstream has fixed this by update i915 audio, hda driver and component API.

Signed-off-by: Liwei Song <liwei.song@windriver.com>
Signed-off-by: chunguang yang <chunguang.yang@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>