ovp: ensure that VFIO is available w/o needing a recompile.
Issue: OVP-2538

There are limitations to PCI pass through, and in some instances
high addresses can get fed to the IOMMU which won't accept anything
with more than 48bits of address.  As per the discussion here:

https://www.marc.info/?t=141300724400001&r=1&w=1

it would seem that VFIO is the current solution of choice, so lets
ensure that we have the module present for our users.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>