Integrate WR specific lttng patches
1/3 [
Author: Jianchuan Wang
Email: jianchuan.wang@windriver.com
Subject: lttng: added lttng Kconfig for in kernel building
Date: Tue, 8 Sep 2015 16:28:56 +0800

Signed-off-by: Paul Barrette<paul.barrette@windriver.com>
Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
]

2/3 [
Author: Jianchuan Wang
Email: jianchuan.wang@windriver.com
Subject: lttng: enable the lttng driver in the build
Date: Tue, 8 Sep 2015 16:30:42 +0800

Enable the build for the lttng module in drivers/stagging/lttng.

Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
]

3/3 [
Author: Linhao Zhang
Email: linhao.zhang@windriver.com
Subject: FIX: lttng build error if !CONFIG_UID16 on arm
Date: Thu, 22 May 2014 09:52:43 +0800

when CONFIG_UID16 is not set, the old_uid_t is not defined.
If the macros such as OVERRIDE_32_sys_getuid16 is not defined,
the type old_uid_t will be used in lttng,
so define the macros to avoid that.

Signed-off-by: Linhao Zhang <linhao.zhang@windriver.com>
Signed-off-by: Jianchuan Wang <jianchuan.wang@windriver.com>
]

Signed-off-by: Yang Shi <yang.shi@windriver.com>