MEI: disable via disabling NFC.
Issue: OVP-2603

This module is blacklisted by several distros since it does bad
things on generic COTS PCs, like slowing bootup via delayed and
failed detection/probing.

Since we've not a hard requirement for NFC, we can simply not
source the NFC config fragment which draws in the MEI stuff.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>