MEI: disable via disabling NFC. (part two)
Issue: OVP-2603

This module is blacklisted by several distros since it does bad
things on generic COTS PCs, like slowing bootup via delayed and
failed detection/probing.

Since we've not a hard requirement for NFC, we can simply not
source the NFC config fragment which draws in the MEI stuff.

We have to do the preempt-rt one as well since we run that scc
in ovp with cfg enabled.  We run the standard.scc with config
disabled so we can leave that remaining instance alone.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>