spi-amd-steppeeagle: Add completion to notify spi_amd_spi
Issue: LIN6-10370

1/1 [
Author: chunguang yang
Email: chunguang.yang@windriver.com
Subject: spi-amd-steppeeagle: Add completion to notify spi_amd_spi
Date: Tue, 29 Sep 2015 14:41:06 +0800

spi-amd-steppeeagle.c use a kernel thread to monitor data transfer
queue. In the thread main loop, it polls the amd_spi->msg_queue to
check if any transfer. Change this so that the polled message queue
won't consume 95% cpu on a core that is running only the polling thread
and reduce power consumption. To fix this, use completion to notify
the thread when data transfer is ready.
Signed-off-by: chunguang yang <chunguang.yang@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>