ipconfig: send Client-identifier in DHCP requests
1/1 [
Author: Li RongQing
Email: roy.qing.li@gmail.com
Subject: ipconfig: send Client-identifier in DHCP requests
Date: Thu, 15 Oct 2015 16:54:36 +0800

commit 26fb342c734061859fec1bd9e987bb6b78061ef0 upstream

A dhcp server may provide parameters to a client from a pool of IP
addresses and using a shared rootfs, or provide a specific set of
parameters for a specific client, usually using the MAC address to
identify each client individually. The dhcp protocol also specifies
a client-id field which can be used to determine the correct
parameters to supply when no MAC address is available. There is
currently no way to tell the kernel to supply a specific client-id,
only the userspace dhcp clients support this feature, but this can
not be used when the network is needed before userspace is available
such as when the root filesystem is on NFS.

This patch is to be able to do something like "ip=dhcp,client_id_type,
client_id_value", as a kernel parameter to enable the kernel to
identify itself to the server.

Signed-off-by: Li RongQing <roy.qing.li@gmail.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
]

Signed-off-by: Yang Shi <yang.shi@windriver.com>