vhost/scsi: potential memory corruption
Issue: LIN7-4739

1/1 [
Author: Dan Carpenter
Email: dan.carpenter@oracle.com
Subject: vhost/scsi: potential memory corruption
Date: Thu, 5 Feb 2015 10:37:33 +0300

commit 59c816c1f24df0204e01851431d3bab3eb76719c upstream

This code in vhost_scsi_make_tpg() is confusing because we limit "tpgt"
to UINT_MAX but the data type of "tpg->tport_tpgt" and that is a u16.

I looked at the context and it turns out that in
vhost_scsi_set_endpoint(), "tpg->tport_tpgt" is used as an offset into
the vs_tpg[] array which has VHOST_SCSI_MAX_TARGET (256) elements so
anything higher than 255 then it is invalid.  I have made that the limit
now.

In vhost_scsi_send_evt() we mask away values higher than 255, but now
that the limit has changed, we don't need the mask.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Nicholas Bellinger <nab@linux-iscsi.org>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>