vmstat: refresh worker commit from mainline to apply to 3.10
Issue: US69814

We refresh to make it applicable to this older baseline. It is
viewed as a separate change to ease review and double checking.

We can see trivial changes around this_cpu_ptr vs get_cpu_ptr
due to the older baseline,  some changes due to the __cpuinit
that I purged from mainline but was still in as of 3.10.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>