jbd2: refine waiting for shadow buffers
Issue: LIN4-33018

1/3 [
Author: Jan Kara
Email: jack@suse.cz
Subject: jbd2: don't create journal_head for temporary journal buffers
Date: Tue, 4 Jun 2013 12:01:45 -0400

commit f5113effc2a2ee6b86a4b345ce557353dcbcfffe upstream

When writing metadata to the journal, we create temporary buffer heads
for that task.  We also attach journal heads to these buffer heads but
the only purpose of the journal heads is to keep buffers linked in
transaction's BJ_IO list.  We remove the need for journal heads by
reusing buffer_head's b_assoc_buffers list for that purpose.  Also
since BJ_IO list is just a temporary list for transaction commit, we
use a private list in jbd2_journal_commit_transaction() for that thus
removing BJ_IO list from transaction completely.

Reviewed-by: Zheng Liu <wenqing.lz@taobao.com>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

2/3 [
Author: Jan Kara
Email: jack@suse.cz
Subject: jbd2: remove journal_head from descriptor buffers
Date: Tue, 4 Jun 2013 12:06:01 -0400

commit e5a120aeb57f40ae568a5ca1dd6ace53d0213582 upstream

Similarly as for metadata buffers, also log descriptor buffers don't
really need the journal head. So strip it and remove BJ_LogCtl list.

Reviewed-by: Zheng Liu <wenqing.lz@taobao.com>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

3/3 [
Author: Jan Kara
Email: jack@suse.cz
Subject: jbd2: refine waiting for shadow buffers
Date: Tue, 4 Jun 2013 12:08:56 -0400

commit b34090e5e22a02fba0e4473056cce9420ad9dd0b upstream

Currently when we add a buffer to a transaction, we wait until the
buffer is removed from BJ_Shadow list (so that we prevent any changes
to the buffer that is just written to the journal).  This can take
unnecessarily long as a lot happens between the time the buffer is
submitted to the journal and the time when we remove the buffer from
BJ_Shadow list.  (e.g.  We wait for all data buffers in the
transaction, we issue a cache flush, etc.)  Also this creates a
dependency of do_get_write_access() on transaction commit (namely
waiting for data IO to complete) which we want to avoid when
implementing transaction reservation.

So we modify commit code to set new BH_Shadow flag when temporary
shadowing buffer is created and we clear that flag once IO on that
buffer is complete.  This allows do_get_write_access() to wait only
for BH_Shadow bit and thus removes the dependency on data IO
completion.

Reviewed-by: Zheng Liu <wenqing.lz@taobao.com>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: Shan Hai <shan.hai@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>