udf: Check path length when reading symlink - linux - CVE-2014-9731
Issue: LIN7-4740

1/1 [
Author: Jan Kara
Email: jack@suse.cz
Subject: udf: Check path length when reading symlink
Date: Thu, 18 Dec 2014 22:37:50 +0100

commit 0e5cc9a40ada6046e6bc3bdfcd0c0d7e4b706b14 upstream

Symlink reading code does not check whether the resulting path fits into
the page provided by the generic code. This isn't as easy as just
checking the symlink size because of various encoding conversions we
perform on path. So we have to check whether there is still enough space
in the buffer on the fly.

CC: stable@vger.kernel.org
Reported-by: Carl Henrik Lunde <chlunde@ping.uio.no>
Signed-off-by: Jan Kara <jack@suse.cz>
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>