drm/i915: Do not disable POWER_DOMAIN_INIT power domain when update crtc
Issue: LIN7-4904

1/1 [
Author: Liwei Song
Email: liwei.song@windriver.com
Subject: drm/i915: Do not disable POWER_DOMAIN_INIT power domain when update crtc
Date: Tue, 17 Nov 2015 16:31:08 +0800

This patch can fix the following issue in baytrail platform:
Boot without VGA connected, after boot it successfully, VGA display
does not become active if plugged in the VGA line.

This is due to the power-domain POWER_DOMAIN_INIT was disabled when
update the crtc, at the same time the DISPLAY power-well which belong
to this power-domain is also disabled, without the enable of DISPLAY,
the interrupt in this power-domain also can not be triggered, so when
connected the VGA, the hotplug event which registed in this interrupt
will not be run.

It can be fixed by stop the crtc update event to disable the
POWER_DOMAIN_INIT domain, and mark this power domain will be used
sometimes by transmiting "true" to the function intel_display_set_init_power.

And Upstream also exist this issue.

Signed-off-by: Liwei Song <liwei.song@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>