Revert "timers: do not raise softirq unconditionally"
Issue: LIN6-10388

1/1 [
Author: "Steven Rostedt (Red Hat)"
Email: rostedt@goodmis.org
Subject: Revert "timers: do not raise softirq unconditionally"
Date: Mon, 2 Nov 2015 17:58:51 +0800

commit faf9582f814d9ac897b2a853d1c06e1d39f0418b in linux-stable-rt.

This reverts commit 891f510568343d93c5aa2f477b6bebe009b48f05.

An issue arisen that if a rt_mutex (spin_lock converted to a mutex
in PREEMPT_RT) is taken in hard interrupt context, it could cause
a false deadlock detection and trigger a BUG_ON() from the return
value of task_blocks_on_rt_mutex() in rt_spin_lock_slowlock().

The problem is this:

    CPU0                        CPU1
    ----                        ----
  spin_lock(A)
                                spin_lock(A)
                        [ blocks, but spins as owner on
                          CPU 0 is running ]

                                <interrupt>
                                        spin_trylock(B)
                                        [ succeeds ]

  spin_lock(B)
  <blocks>

Now the deadlock detection triggers and follows the locking:

  Task X (on CPU0) blocked on spinlock B owned by task Y on
  CPU1 (via the interrupt taking it with a try lock)

  The owner of B (Y) is blocked on spin_lock A (still spinning)
  A is owned by task X (self). DEADLOCK detected! BUG_ON triggered.

This was caused by the code to try to not raise softirq unconditionally
to allow NO_HZ_FULL to work. Unfortunately, reverting that patch causes
NO_HZ_FULL to break again, but that's still better than triggering
a BUG_ON().

Signed-off-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Hu <yadi.hu@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>