random: use the architectural HWRNG for the SHA's IV in extract_buf()
Issue: LIN6-10531

1/3 [
Author: yzhu1
Email: yzhu1@windriver.com
Subject: Revert "random: use the architectural HWRNG for the SHA's IV in extract_buf()"
Date: Tue, 24 Nov 2015 17:13:53 +0800

The backport of the commit 8f494e8 "random: use the architectural HWRNG for
the SHA's IV in extract_buf()" was incomplete. And we are reverting it
to start over with a new set of changes.

Signed-off-by: yzhu1 <yzhu1@windriver.com>
]

2/3 [
Author: Theodore Ts'o
Email: tytso@mit.edu
Subject: random: mix in architectural randomness earlier in extract_buf()
Date: Wed, 25 Nov 2015 18:41:42 +0800

commit 85a1f77716cf546d9b9c42e2848b5712f51ba1ee upstream

Previously if CPU chip had a built-in random number generator (i.e.,
RDRAND on newer x86 chips), we mixed it in at the very end of
extract_buf() using an XOR operation.

We now mix it in right after the calculate a hash across the entire
pool.  This has the advantage that any contribution of entropy from
the CPU's HWRNG will get mixed back into the pool.  In addition, it
means that if the HWRNG has any defects (either accidentally or
maliciously introduced), this will be mitigated via the non-linear
transform of the SHA-1 hash function before we hand out generated
output.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: yzhu1 <yzhu1@windriver.com>
]

3/3 [
Author: Theodore Ts'o
Email: tytso@mit.edu
Subject: random: use the architectural HWRNG for the SHA's IV in extract_buf()
Date: Wed, 25 Nov 2015 19:00:54 +0800

commit 46884442fc5bb81a896f7245bd850fde9b435509 upstream

To help assuage the fears of those who think the NSA can introduce a
massive hack into the instruction decode and out of order execution
engine in the CPU without hundreds of Intel engineers knowing about
it (only one of which woud need to have the conscience and courage of
Edward Snowden to spill the beans to the public), use the HWRNG to
initialize the SHA starting value, instead of xor'ing it in
afterwards.

Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>
Signed-off-by: yzhu1 <yzhu1@windriver.com>
]

Signed-off-by: Yue Tao <yue.tao@windriver.com>