KVM: x86: PREFETCH and HINT_NOP should have SrcMem flag - linux - CVE-2014-8480
Issue: LIN7-2168

1/1 [
Author: Nadav Amit
Email: namit@cs.technion.ac.il
Subject: KVM: x86: PREFETCH and HINT_NOP should have SrcMem flag
Date: Mon, 30 Nov 2015 16:44:16 +0800

The decode phase of the x86 emulator assumes that every instruction with
the ModRM flag, and which can be used with RIP-relative addressing, has
either SrcMem or DstMem.  This is not the case for several instructions -
prefetch, hint-nop and clflush.

Adding SrcMem|NoAccess for prefetch and hint-nop and SrcMem for clflush.

This fixes CVE-2014-8480.

Signed-off-by: Nadav Amit <namit@cs.technion.ac.il>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Hu <yadi.hu@windriver.com>
]

Signed-off-by: Yue Tao <Yue.Tao@windriver.com>