fs/block_dev.c: Remove WARN_ON() when inode writeback fails
1/1 [
Author: Vivek Goyal
Email: vgoyal@redhat.com
Subject: fs/block_dev.c: Remove WARN_ON() when inode writeback fails
Date: Mon, 9 Nov 2015 09:23:40 -0700

commit dbd3ca50753e70e09cad747dce23b1a7683a3342 upstream.

If a block device is hot removed and later last reference to device
is put, we try to writeback the dirty inode. But device is gone and
that writeback fails.

Currently we do a WARN_ON() which does not seem to be the right thing.
Convert it to a ratelimited kernel warning.

Reported-by: Andi Kleen <andi@firstfloor.org>
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Acked-by: Tejun Heo <tj@kernel.org>
[jmoyer@redhat.com: get rid of unnecessary name initialization, 80 cols]
Signed-off-by: Jeff Moyer <jmoyer@redhat.com>
Signed-off-by: Jens Axboe <axboe@fb.com>

Signed-off-by: Liwei Song <liwei.song@windriver.com>
]

Signed-off-by: Yang Shi <yang.shi@windriver.com>