pstore: create pstore config
Rally: US72885

on platforms when there's a persistent memory storage (or when ram is
not cleared during bootup like on arm), pstore and it's pstore-ram
backend can be used to save log messages on a previously defined
and unallocated memory portion. and if bootup messages are needed to
save, the pstore-ram backend must be build in.

and enable it by default on cgl ktype

Signed-off-by: Roy Li <rongqing.li@windriver.com>